move schema into configs
